PolygonImage is Replaced with PolygonSprite so that we can use SHADERs
PolygonImage.cs is no more used in the game as it uses RAWImage component which accepts the texture type image.
but Now we use PolygonSprite prefab in GameArea.cs which uses PolygonSprite.cs script.
PolygonSprite.cs uses the IMAGE component in the Inspector which allows us to have SPRITEs and using SPRITEs we can assign shaders to have visual Effects